+++ mthree/mitigation.py
@@ -31,10 +31,11 @@ from mthree.circuits import (
     balanced_cal_strings,
     balanced_cal_circuits,
 )
-from mthree.matrix import _reduced_cal_matrix, sdd_check
+from mthree.src.matrix import _reduced_cal_matrix, sdd_check
+from mthree.src.matvec import M3MatVec
+
 from mthree.utils import counts_to_vector, vector_to_quasiprobs
 from mthree.norms import ainv_onenorm_est_lu, ainv_onenorm_est_iter
-from mthree.matvec import M3MatVec
 from mthree.exceptions import M3Error
 from mthree.classes import QuasiCollection
 from ._helpers import system_info
