+++ mthree/src/compute.pxd
