+++ mthree/src/compute.pyx
