+++ mthree/src/expval.pyx
