+++ mthree/src/hamming.pyx
