+++ mthree/src/matrix.pyx
