+++ mthree/src/matvec.pyx
