+++ mthree/src/probability.pyx
